requirements.txt
PyYAML==3.12
jieba==0.39
termcolor==1.1.0
bs4==0.0.1
six==1.11.0
termcolor==1.1.0
terminaltables==3.1.0
opencv-python==3.3.0.10
numpy==1.13.1
